reduced header from mega component into the parent of 4 children
